requirements.txt
langchain 
langchain-community 
langchain-experimental
llama-index-embeddings-huggingface
llama-cpp-python
llama-index-experimental
llama-index
llama-index-llms-langchain
python-dotenv
flask
loguru 
accelerate